Updated SIDIA tests
I got rid of tests that weren't being called and tests for APIs that are
now gone. I changed usage of GVRScript to GVRMain. I updated collider
API usage in some places. Now all the tests pass except those that
create shaders. There appears to be a problem with the test framework
and our framework involving when objects are created. Synchronization
might fix the problem.