Added bindShaders calls, access content from network
The jassimp assets were moved to asset-tests but I forgot that the
TextureTests in framework-tests used them too. I changed the tests to
access the files from github instead of the assets directory. I also
added missing bindShaders calls.

As to why we have to add bindShaders calls now, it is because I fixed a
bug where we were using lit shaders when there were no light sources in
the scene (gvrf should use unlit shaders in this case). Now you need to
call bindShaders after the light sources are added to get lit shaders.
In a future version of GVRF I will change the renderer to spawn
bindShaders automatically on the Java thread.