Golden masters for S7 Edge (#42)
* Revert "adding beach.zip for asset-tests"

This reverts commit 32b58e0b59e70b3a432f41f81f255e008d0d9380.

* Adding golden masters for S7 Edge

Since GearVRF doesn't work on the Note4 anymore, I have regenerated the golden master images for the asset-tests and framework-tests. I have a new directory, golden_masters, which contains the images for both Note4 and S7Edge. The actual assets directories of the test folders only contains the S7Edge images.
Mihail mentioned he would write scripts to copy the images to the SD card. Maybe we can determine what phone it is and copy the right ones? I am not sure if this is even possible but it would be nice if we could.