Golden Masters automatically selected based on phone model.
GearVRf-Tests/golden_masters directory in the GitHub repository has now 2 kings of objects:
1. Directories with golden master images, named generically: "S7Euro", "Note4"
2. Text files named with exact phone model "SM-G935F" containing text string redirecting to the directory: "S7Edge"
        (symbolic links will also work, but are not recommended)

If phone model you are testing is not present, look for error message in the Android log:
"Golden master redirect not found: " with the exact URL of the text file.
Create this text file containing the directory name,
and, if necessary, create and populate golden master directory for your phone.

After this error, test will continue with the default directory, currently "S7Edge"