Adding GLTF tests (#266)
* Adding GLTF tests

Adding 3 new tests for GLTF import and their screenshots.
Also adding missing screenshots for the SIDIA render config and render shader tests.

The GLTF test with embedded textures doesn't work right now. I will investigate and come up with a fix if possible. The other 2 tests work.

This PR depends on PR #1842 in GearVRf.

* replaced bad texture in GLTF tests

The texture CesiumLogoFlat.png used by the GLTF tests was a corrupted file. I replaced it with the GearVRF logo and redid the screen shots.