Fix render tests, add more gltf2 tests (#271)
* Fix render tests, add more gltf2 tests

RenderShaderTests had bad golden masters because the test did not wait
for textures to load and failed with black screen intermittently. Some
tests had textured materials with no texcoord - also black screen. I
fixed the tests to wait for texture loading and skip the cases that
don't work. Now they all pass.

Fixed RenderConfig rendering order test

Added picking test for GVRColliderGroup

* fix texture count in custom IO test

The implementation of assimp IO changed with 4.1. It used to open the resource twice, once to check if it exists and then to read it. Now it opens the resource once.